+++ hastexo/hastexo.py
@@ -110,6 +110,12 @@ class HastexoXBlock(XBlock,
         help="Set the progress check button label. "
              "For example: \"Submit Answer\" or \"Check Progress\"(Default)."
     )
+    show_hints_on_error = Boolean(
+        default=True,
+        scope=Scope.settings,
+        help="On progress check failure, display the tests' standard error "
+             "streams as hints"
+    )
 
     # Set via XML
     hook_events = Dict(
@@ -169,6 +175,7 @@ class HastexoXBlock(XBlock,
     editable_fields = (
         'display_name',
         'progress_check_label',
+        'show_hints_on_error',
         'weight',
         'stack_template_path',
         'hook_script',
@@ -379,6 +386,7 @@ class HastexoXBlock(XBlock,
         node.set('xblock-family', self.entry_point)
         node.set('display_name', self.display_name)
         node.set('progress_check_label', self.progress_check_label)
+        node.set('show_hints_on_error', str(self.show_hints_on_error))
         node.set('weight', str(self.weight))
         node.set('stack_user_name', self.stack_user_name)
         node.set('stack_protocol', self.stack_protocol)
@@ -528,7 +536,8 @@ class HastexoXBlock(XBlock,
             "font_size": settings.get("terminal_font_size"),
             "instructions_layout": settings.get("instructions_layout"),
             "read_only": self.read_only,
-            "progress_check_label": self.progress_check_label
+            "progress_check_label": self.progress_check_label,
+            "show_hints_on_error": self.show_hints_on_error
         })
 
         return frag