+++ Changelog.md
@@ -1,3 +1,7 @@
+Unreleased
+-------------------------
+[Enhancement] Make the progress check button label configurable.
+
 Version 4.1.5 (2021-01-25)
 --------------------------
 * [Bug fix] Fix ICMP connectivity check for IPv6-only stacks.